refactor: 迁移单Agent执行引擎至 agent_runtime/ 顶层共享层
