+++ src/hivememory/agent_runtime/koakuma.py
@@ -40,10 +40,10 @@ from hivememory.core.mtp import (
     MTPParseError,
     MTPFormatter,
 )
-from hivememory.alice.runtime.cache import KoakumaAtomCache
-from hivememory.alice.runtime.models import MTPExecutionContext
-from hivememory.alice.runtime.pending_atom import PendingAtomRuntime
-from hivememory.alice.runtime.resolver import RuntimeAliasResolver
+from hivememory.agent_runtime.cache import KoakumaAtomCache
+from hivememory.agent_runtime.models import MTPExecutionContext
+from hivememory.agent_runtime.pending_atom import PendingAtomRuntime
+from hivememory.agent_runtime.resolver import RuntimeAliasResolver
 from hivememory.engines.memory_compiler import MemoryCompiler, MemoryCompileTarget, MemoryCompileOptions
 from hivememory.core.protocol.models import (
     MTPExecutionResult,
@@ -61,7 +61,7 @@ from hivememory.core.mtp.exceptions import (
 )
 if TYPE_CHECKING:
     from hivememory.alice.runtime.bus import AliceBus
-    from hivememory.alice.runtime.resolver import ResolveResult
+    from hivememory.agent_runtime.resolver import ResolveResult
     from hivememory.system.config import KoakumaConfig
 
 logger = logging.getLogger(__name__)
@@ -118,7 +118,7 @@ class KoakumaRuntime:
 
         # 初始化内核工具注册表 KERNEL_REGISTRY (Section 4.2.1)
         # 硬编码的 sys_ 工具集，随系统启动加载，Zero Latency
-        from hivememory.alice.runtime.syscalls import build_kernel_registry
+        from hivememory.agent_runtime.syscalls import build_kernel_registry
         self._kernel_registry = build_kernel_registry(
             python_repl_timeout=self._config.python_repl_timeout_seconds,
             workspace_path=self._config.workspace_path,
@@ -907,7 +907,7 @@ class KoakumaRuntime:
         Returns:
             MTPResponse: 执行结果
         """
-        from hivememory.alice.runtime.syscalls import execute_sandboxed
+        from hivememory.agent_runtime.syscalls import execute_sandboxed
 
         result = execute_sandboxed(
             code,
