+++ src/hivememory/agent_runtime/syscalls/file_io.py
