+++ src/hivememory/agent_runtime/syscalls/repl.py
