+++ src/hivememory/agent_runtime/syscalls/types.py
