+++ src/hivememory/agent_runtime/syscalls/web_search.py
