+++ src/hivememory/alice/runtime/agent/__init__.py
@@ -1,21 +1,11 @@
-"""Agent runtime component exports."""
+"""Alice agent orchestration component exports."""
 
 from hivememory.alice.runtime.agent.frame_scheduler import FrameScheduler
-from hivememory.alice.runtime.agent.loop_executor import KernelLoopExecutor
-from hivememory.alice.runtime.agent.mtp_executor import (
-    KoakumaMTPExecutor,
-    MTPExecutor,
-)
 from hivememory.alice.runtime.agent.profile_resolver import AgentProfileResolver
 from hivememory.alice.runtime.agent.runtime import AgentRuntime
-from hivememory.alice.runtime.agent.worker_agent import WorkerAgentService
 
 __all__ = [
     "AgentProfileResolver",
     "AgentRuntime",
     "FrameScheduler",
-    "KernelLoopExecutor",
-    "KoakumaMTPExecutor",
-    "MTPExecutor",
-    "WorkerAgentService",
 ]
