+++ src/hivememory/alice/runtime/orchestrator.py
@@ -17,7 +17,7 @@ import asyncio
 import logging
 from typing import TYPE_CHECKING, Any, AsyncGenerator, Callable, Dict, List, Optional
 
-from hivememory.alice.runtime.models import (
+from hivememory.agent_runtime.models import (
     ExecutionFrame,
     FrameExecutionResult,
     FrameExecutionStatus,
@@ -35,10 +35,10 @@ from hivememory.engines.memory_compiler import (
 
 if TYPE_CHECKING:
     from hivememory.alice.runtime.agent.frame_scheduler import FrameScheduler
-    from hivememory.alice.runtime.agent.loop_executor import KernelLoopExecutor
+    from hivememory.agent_runtime.loop_executor import KernelLoopExecutor
     from hivememory.alice.runtime.agent.profile_resolver import AgentProfileResolver
-    from hivememory.alice.runtime.pending_atom import PendingAtomRuntime
-    from hivememory.alice.runtime.resolver import RuntimeAliasResolver
+    from hivememory.agent_runtime.pending_atom import PendingAtomRuntime
+    from hivememory.agent_runtime.resolver import RuntimeAliasResolver
     from hivememory.core.models import AgentProfile, Identity
 
 logger = logging.getLogger(__name__)
