+++ tests/e2e/pipeline/test_kernel_loop_e2e.py
@@ -27,8 +27,8 @@ import pytest
 from unittest.mock import MagicMock
 
 from hivememory.system.config import LLMConfig, KoakumaConfig
-from hivememory.alice.runtime.agent.worker_agent import WorkerAgentService
-from hivememory.alice.runtime.koakuma import KoakumaRuntime
+from hivememory.agent_runtime.worker_agent import WorkerAgentService
+from hivememory.agent_runtime.koakuma import KoakumaRuntime
 from hivememory.core.protocol.models import AgentRunResult, RetrievalResponse
 from hivememory.prompts.mtp import MTPPromptBuilder
 from hivememory.patchouli.system import PatchouliSystem
