+++ tests/unit/alice/runtime/syscalls/live_support.py
@@ -7,8 +7,8 @@ from typing import Dict, List, Optional, Tuple
 from unittest.mock import MagicMock
 
 from hivememory.system.config import LLMConfig, KoakumaConfig
-from hivememory.alice.runtime.koakuma import KoakumaRuntime
-from hivememory.alice.runtime.models import MTPExecutionContext
+from hivememory.agent_runtime.koakuma import KoakumaRuntime
+from hivememory.agent_runtime.models import MTPExecutionContext
 from hivememory.core.mtp import (
     MTP_LEFT_DELIMITER,
     MTP_RIGHT_DELIMITER,
