+++ tests/unit/alice/runtime/syscalls/test_clock.py
@@ -1,6 +1,6 @@
 import re
 
-from hivememory.alice.runtime.syscalls.clock import sys_clock
+from hivememory.agent_runtime.syscalls.clock import sys_clock
 
 
 class TestSysClock:
