+++ tests/unit/engines/memory_compiler/test_compiler.py
@@ -358,7 +358,7 @@ class TestResolveResultCompilation:
 
     @pytest.fixture
     def atom_resolve(self, sample_atom):
-        from hivememory.alice.runtime.resolver import ResolveResult
+        from hivememory.agent_runtime.resolver import ResolveResult
 
         return ResolveResult(
             kind="atom",
@@ -368,7 +368,7 @@ class TestResolveResultCompilation:
 
     @pytest.fixture
     def redirect_resolve(self, sample_atom):
-        from hivememory.alice.runtime.resolver import ResolveResult
+        from hivememory.agent_runtime.resolver import ResolveResult
         from hivememory.core.models import (
             PendingAtomResolution,
             PendingAtomSettlement,
@@ -389,7 +389,7 @@ class TestResolveResultCompilation:
 
     @pytest.fixture
     def discarded_resolve(self):
-        from hivememory.alice.runtime.resolver import ResolveResult
+        from hivememory.agent_runtime.resolver import ResolveResult
         from hivememory.core.models import (
             PendingAtomResolution,
             PendingAtomSettlement,
@@ -409,7 +409,7 @@ class TestResolveResultCompilation:
 
     @pytest.fixture
     def pending_resolve(self):
-        from hivememory.alice.runtime.resolver import ResolveResult
+        from hivememory.agent_runtime.resolver import ResolveResult
         from hivememory.core.models import PendingAtom, PendingAtomStatus, WriteFocus
 
         pending = PendingAtom(
@@ -464,7 +464,7 @@ class TestResolveResultCompilation:
         assert "Pending content" in artifact.text
 
     def test_not_found_raises(self, compiler):
-        from hivememory.alice.runtime.resolver import ResolveResult
+        from hivememory.agent_runtime.resolver import ResolveResult
 
         resolve = ResolveResult(kind="not_found", requested_alias="missing")
         with pytest.raises(ValueError, match="not_found"):
