+++ tests/unit/patchouli/test_worker_agent.py
@@ -10,7 +10,7 @@ WorkerAgentService 单元测试
 import pytest
 from unittest.mock import Mock, patch
 
-from hivememory.alice.runtime.agent.worker_agent import WorkerAgentService
+from hivememory.agent_runtime.worker_agent import WorkerAgentService
 from hivememory.system.config import LLMConfig
 from hivememory.core.mtp import MTP_LEFT_DELIMITER, MTP_STOP_SEQUENCE
 
@@ -41,7 +41,7 @@ class TestWorkerAgentGenerateAsync:
         self.config = _make_config()
         self.service = WorkerAgentService(config=self.config)
 
-    @patch("hivememory.alice.runtime.agent.worker_agent.litellm.acompletion")
+    @patch("hivememory.agent_runtime.worker_agent.litellm.acompletion")
     async def test_normal_completion(self, mock_completion):
         """正常完成，无 MTP 中断"""
         mock_completion.return_value = _make_response("普通回复", "stop")
@@ -56,7 +56,7 @@ class TestWorkerAgentGenerateAsync:
         assert result.prefix_text == "普通回复"
         assert result.mtp_fragment == ""
 
-    @patch("hivememory.alice.runtime.agent.worker_agent.litellm.acompletion")
+    @patch("hivememory.agent_runtime.worker_agent.litellm.acompletion")
     async def test_mtp_interrupted(self, mock_completion):
         """finish_reason=stop 且文本含 ⟪ 时检测为 MTP 中断"""
         text = f"前面的文本{MTP_LEFT_DELIMITER}READ|mem_doc"
@@ -70,7 +70,7 @@ class TestWorkerAgentGenerateAsync:
         assert result.prefix_text == "前面的文本"
         assert result.mtp_fragment == f"{MTP_LEFT_DELIMITER}READ|mem_doc"
 
-    @patch("hivememory.alice.runtime.agent.worker_agent.litellm.acompletion")
+    @patch("hivememory.agent_runtime.worker_agent.litellm.acompletion")
     async def test_length_limited_not_mtp(self, mock_completion):
         """finish_reason=length 时即使含 ⟪ 也不标记为 MTP"""
         text = f"文本{MTP_LEFT_DELIMITER}something"
@@ -84,7 +84,7 @@ class TestWorkerAgentGenerateAsync:
         assert result.prefix_text == text
         assert result.mtp_fragment == ""
 
-    @patch("hivememory.alice.runtime.agent.worker_agent.litellm.acompletion")
+    @patch("hivememory.agent_runtime.worker_agent.litellm.acompletion")
     async def test_empty_text_response(self, mock_completion):
         """LLM 返回空文本"""
         mock_completion.return_value = _make_response("", "stop")
@@ -96,7 +96,7 @@ class TestWorkerAgentGenerateAsync:
         assert result.text == ""
         assert result.was_mtp_interrupted is False
 
-    @patch("hivememory.alice.runtime.agent.worker_agent.litellm.acompletion")
+    @patch("hivememory.agent_runtime.worker_agent.litellm.acompletion")
     async def test_mtp_delimiter_at_start(self, mock_completion):
         """⟪ 在位置 0，prefix_text 为空"""
         text = f"{MTP_LEFT_DELIMITER}SEARCH|query=\"test\""
@@ -110,7 +110,7 @@ class TestWorkerAgentGenerateAsync:
         assert result.prefix_text == ""
         assert result.mtp_fragment == text
 
-    @patch("hivememory.alice.runtime.agent.worker_agent.litellm.acompletion")
+    @patch("hivememory.agent_runtime.worker_agent.litellm.acompletion")
     async def test_multiple_delimiters_uses_last(self, mock_completion):
         """多个 ⟪ 使用 rfind 取最后一个"""
         text = f"a{MTP_LEFT_DELIMITER}first{MTP_LEFT_DELIMITER}second"
@@ -125,7 +125,7 @@ class TestWorkerAgentGenerateAsync:
         assert result.prefix_text == text[:last_pos]
         assert result.mtp_fragment == text[last_pos:]
 
-    @patch("hivememory.alice.runtime.agent.worker_agent.litellm.acompletion")
+    @patch("hivememory.agent_runtime.worker_agent.litellm.acompletion")
     async def test_llm_exception_propagated(self, mock_completion):
         """litellm.acompletion 抛异常时向上传播"""
         mock_completion.side_effect = RuntimeError("API error")
@@ -135,7 +135,7 @@ class TestWorkerAgentGenerateAsync:
                 [{"role": "user", "content": "test"}]
             )
 
-    @patch("hivememory.alice.runtime.agent.worker_agent.litellm.acompletion")
+    @patch("hivememory.agent_runtime.worker_agent.litellm.acompletion")
     async def test_stop_sequence_injected(self, mock_completion):
         """验证 stop=[MTP_STOP_SEQUENCE] 被传入"""
         mock_completion.return_value = _make_response("ok", "stop")
@@ -145,7 +145,7 @@ class TestWorkerAgentGenerateAsync:
         call_kwargs = mock_completion.call_args
         assert call_kwargs[1]["stop"] == [MTP_STOP_SEQUENCE]
 
-    @patch("hivememory.alice.runtime.agent.worker_agent.litellm.acompletion")
+    @patch("hivememory.agent_runtime.worker_agent.litellm.acompletion")
     async def test_kwargs_passed_through(self, mock_completion):
         """额外 kwargs 正确传递给 litellm"""
         mock_completion.return_value = _make_response("ok", "stop")
@@ -188,7 +188,7 @@ class TestWorkerAgentGenerateStream:
         self.config = _make_config()
         self.service = WorkerAgentService(config=self.config)
 
-    @patch("hivememory.alice.runtime.agent.worker_agent.litellm.acompletion")
+    @patch("hivememory.agent_runtime.worker_agent.litellm.acompletion")
     async def test_stream_no_duplicate_before_mtp(self, mock_completion):
         left = MTP_LEFT_DELIMITER
         chunks = [
@@ -210,7 +210,7 @@ class TestWorkerAgentGenerateStream:
         assert stream_chunks[-1].result is not None
         assert stream_chunks[-1].result.was_mtp_interrupted is True
 
-    @patch("hivememory.alice.runtime.agent.worker_agent.litellm.acompletion")
+    @patch("hivememory.agent_runtime.worker_agent.litellm.acompletion")
     async def test_stream_flushes_pending_tail_without_mtp(self, mock_completion):
         chunks = [
             _make_stream_chunk("A"),
